git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@9091 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/enhance.c
@@ -2411,7 +2411,7 @@ MagickExport MagickBooleanType HaldClutImageChannel(Image *image,
       point.x=QuantumScale*(level-1.0)*GetPixelRed(q);
       point.y=QuantumScale*(level-1.0)*GetPixelGreen(q);
       point.z=QuantumScale*(level-1.0)*GetPixelBlue(q);
-      offset=point.x+level*floor(point.y)+cube_size*floor(point.z);
+      offset=(double) (point.x+level*floor(point.y)+cube_size*floor(point.z));
       point.x-=floor(point.x);
       point.y-=floor(point.y);
       point.z-=floor(point.z);
